assignment3_stock_graph/pyproject.toml
[project]
name = "assignment3-stock-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpha-vantage==2.3.1",
    "fastapi==0.104.1",
    "finnhub-python==2.4.18",
    "matplotlib==3.8.2",
    "newsapi-python==0.2.7",
    "pandas==2.1.3",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "uvicorn==0.24.0",
    "yfinance==0.2.31",
]
